Refactor: move Express setup to dedicated loader function
