+++ package.json
@@ -7,12 +7,12 @@
   "engines": {
     "node": ">=16.0.0"
   },
-  "main": "src/server.ts",
+  "main": "src/infra/http/server.ts",
   "scripts": {
     "build": "tsc && tsc-alias",
-    "dev": "NODE_ENV=development ts-node-dev -r tsconfig-paths/register ./src/server.ts",
-    "prod": "NODE_ENV=production ts-node -r tsconfig-paths/register ./src/server.ts",
-    "start": "NODE_ENV=production node dist/server.js",
+    "dev": "NODE_ENV=development ts-node-dev -r tsconfig-paths/register ./src/infra/http/server.ts",
+    "prod": "NODE_ENV=production ts-node -r tsconfig-paths/register ./src/infra/http/server.ts",
+    "start": "NODE_ENV=production node dist/infra/http/server.js",
     "test": "NODE_ENV=test jest --coverage --verbose",
     "test:ci": "NODE_ENV=testci jest --config ./jest.config.js --collectCoverage --coverageDirectory=\"./coverage\" --ci --reporters=default --reporters=jest-junit --watchAll=false --verbose --silent=false"
   },
