+++ src/infra/http/expressLoader.ts
@@ -1,0 +1,45 @@
+import { Application, Request, Response } from "express";
+import bodyParser, { urlencoded } from "body-parser";
+import { pinoHttp } from "pino-http";
+import { appConfig } from "@utils/appConfig";
+import logger from "@utils/logger";
+
+const loadExpress = async ({ app }: { app: Application }) => {
+  app.use(bodyParser.json());
+  app.use(urlencoded({ extended: true }));
+
+  app.use(pinoHttp({ logger }));
+
+  /* Add CORS headers to all responses. */
+  app.use((req, res, next) => {
+    res.setHeader(
+      "Access-Control-Allow-Origin",
+      appConfig.express.corsOrigin as string,
+    );
+    res.setHeader("Access-Control-Allow-Credentials", "true");
+    res.setHeader(
+      "Access-Control-Allow-Headers",
+      "Origin, X-Requested-With, Content-Type, Accept, Authorization",
+    );
+    res.setHeader("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE");
+    next();
+  });
+
+  app.use("/", (req: Request, res: Response) => {
+    logger.info("Received request...");
+    return res.json({ message: "Hello World!" });
+  });
+
+  app.use("/users", (req: Request, res: Response) => {
+    const { emailAddress, username } = req.body;
+
+    return res.json({
+      user: {
+        emailAddress,
+        username,
+      },
+    });
+  });
+};
+
+export default loadExpress;
