--- src/server.test.ts
@@ -1,5 +1,0 @@
-describe("An example test", () => {
-  it("true should be true", () => {
-    expect(true).toBe(true);
-  });
-});
