+++ tsconfig.json
@@ -34,6 +34,10 @@
     "paths": {
       "@utils": ["utils"],
       "@utils/*": ["utils/*"],
+      "@components": ["components"],
+      "@components/*": ["components/*"],
+      "@infra": ["infra"],
+      "@infra/*": ["infra/*"],
       "@/*": ["./*"]
     } /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
